+++ CHANGELOG.md
@@ -6,6 +6,7 @@ PowerModels.jl Change Log
 - Cleaned up generator data in Matpower files
 - Cleaned up generator data in PSS(R)E PTI file parse
 - Fix source_id in PSS(R)E PTI parse of buses to be only bus id
+- Fix small PSS(R)E PTI parser errors (unicode handling, reserved characters in comments, error message handling)
 
 ### v0.9.2
 - Added tracking of modifications in check_network_data